+++ lib/swiss_knife_vm.dart
@@ -1,0 +1,10 @@
+library swiss_knife;
+
+export 'src/events.dart';
+export 'src/math.dart';
+export 'src/date.dart';
+export 'src/collections.dart';
+export 'src/utils.dart';
+export 'src/data.dart';
+export 'src/io.dart';
+